Limit hunt to spec & test directories if present
If your project has a lot of files, using findfile() with a path of '**'
can take quite some time. We can optimise this by restricting the path
if you're specifically search for a spec or a test file and there is a
corresponding 'spec' or 'test' directory.

If there is no 'spec' or 'test' directory, fallback to searching '**'
again but note that this may change behaviour if you store tests in
different locations and have directories called 'spec' or 'test'.