Look up features even if there are specs
Previously, if a project had specs _and_ features (as is common), it
wasn't possible to trigger Cucumber from the feature step definitions as
it would only check the spec directory for matches. To fix this, still
try the spec and test directories but fall through to the features if no
match is found.